fix determination of version number when building from an sdist archive

+++ setup.py
@@ -1,25 +1,30 @@
 #!/bin/env python
 
+from os import path
 from setuptools import setup
 from subprocess import Popen, PIPE
 
-# write the git version to substratestack/version.py
-# based on version.py by Douglas Creager <dcreager@dcreager.net>
-# http://dcreager.net/2010/02/10/setuptools-git-version-numbers/
-try:
-    p = Popen(['git', 'describe', '--abbrev=4'],
-              stdout=PIPE, stderr=PIPE)
-    p.stderr.close()
-    line = p.stdout.readlines()[0]
-    version = line.strip()[1:]
-except:
-    print("A problem occured while trying to run git. "
-          "Version information is unavailable!")
-    version = 'unknown'
+# check whether we're installing from an sdist archive or not
+if path.exists('PKG-INFO'):
+    from substratestack import __version__ as version
+else:
+    # write the git version to substratestack/version.py
+    # based on version.py by Douglas Creager <dcreager@dcreager.net>
+    # http://dcreager.net/2010/02/10/setuptools-git-version-numbers/
+    try:
+        p = Popen(['git', 'describe', '--abbrev=4'],
+                  stdout=PIPE, stderr=PIPE)
+        p.stderr.close()
+        line = p.stdout.readlines()[0]
+        version = line.strip()[1:]
+    except:
+        print("A problem occured while trying to run git. "
+              "Version information is unavailable!")
+        version = 'unknown'
 
-version_file = open('substratestack/version.py', 'w')
-version_file.write("__version__ = '%s'\n" % version)
-version_file.close()
+    version_file = open('substratestack/version.py', 'w')
+    version_file.write("__version__ = '%s'\n" % version)
+    version_file.close()
 
 
 setup(
